Create Processing 3 code
To write Processing 3 code for an Arduino radar system, you'll need to import libraries and set up serial communication. Create a 3D shape to display the radar screen and use trigonometry to map detected objects. Control the servo motor based on sensor data and optimize the code for performance.To write Processing 3 code for an Arduino radar system, you'll need to import libraries and set up serial communication. Create a 3D shape to display the radar screen and use trigonometry to map detected objects. Control the servo motor based on sensor data and optimize the code for performance.