Update README.md
An Arduino radar system is a device that uses a combination of ultrasonic sensors and servo motors to detect and track objects in its vicinity. It works by emitting high frequency sound waves that bounce off objects and return to the sensor, which then calculates the distance and angle of the object. The servo motors are used to rotate the sensor and provide a 360-degree view of the surroundings. This system can be used for a variety of applications, including robotics, security systems, and autonomous vehicles.